+++ CHANGELOG.md
@@ -3,6 +3,10 @@
 ## unreleased -- v0.3.1
 ### Added
 - Link Details endpoint
+- Redis cache check for cover process
+### Fixed
+- HathiTrust cover process error handling
+- Clustering process on very large groups of records
 
 ## 2021-03-05 -- v0.3.0
 ### Added